+++ persistence/src/main/scala/com/codefactory/ecommerce/tableQuerry/Product/ProductDao.scala
@@ -4,12 +4,15 @@ import slick.jdbc.MySQLProfile.backend
 
 trait ProductDao[T, U[_]] {
 
-//  def sortProductByType(): U[Seq[Int, String, Int, Int, String]]
+  def sortProductByTypeAsc()(implicit db: backend.Database): U[Seq[T]]
 
   def selectProductById(id: Int)(implicit db: backend.Database): U[Option[T]]
 
-  def updateProducts(id: Int, row: T)(implicit db: backend.Database): U[Option[T]]
+  def updateProducts(id: Int, row: T)(
+      implicit db: backend.Database): U[Option[T]]
 
   def insertProduct(row: T)(implicit db: backend.Database): U[T]
+
+  def sortProductByTypeDesc()(implicit db: backend.Database): U[Seq[T]]
 
 }